Modularize code and make ephemeral key exchange and encryption work
Split logic into packages on the server

Fixes errors on AES-GCM and RSA encryption on client side

Signs ephemeral key response from server with RSA private key and
decrypt with the public key on the client

Don't decrypt messages on the server

Replace old logic of curve algorithms to use crypto/ecdh, include logic
of converting the keys to a jwk format accepted by the frontend

TODO: Messages are being sent twice on the htmx:wsConfigSend event
handler. ClientID is not being considered and the websocket response is
a json string instead of a normal json. Signature verification